refactor for mobile view;  separate out DashboardMain into its own component for mobile view
