+++ src/Routes/DashboardRoute/DashboardRoute.css
@@ -2,11 +2,8 @@
   display: flex;
   padding-top: 10px;
 }
+
 .Dashboard_main {
-  flex:3;
-}
-.Dashboard__games-list,
-.Dashboard__game-squads-list {
   flex: 3;
 }
 
@@ -15,9 +12,9 @@
 }
 
 .Dashboard__user-access h3 {
-  margin:0;
-  color:var(--black);
-  padding:22px 0px;
+  margin: 0;
+  color: var(--black);
+  padding: 22px 0px;
   font-size: 28px;
   text-align: center;
   border-bottom: 1px solid var(--greyLines);
@@ -25,23 +22,23 @@
 
 .Dashboard__user-squads,
 .Dashboard__chat {
-  flex:1;
+  flex: 1;
 }
 
 .Dashboard__user-squads {
   margin-top: 0;
 }
 
-.Dashboard_squad-list{
+.Dashboard_squad-list {
   border-right: 1px solid var(--greyLines);
-  flex:1;
+  flex: 1;
 }
 
-.Chat-SquadList-Container{
+.Chat-SquadList-Container {
   text-align: center;
 }
 
-#mobile-buttons-container{
+#mobile-buttons-container {
   position: fixed;
   width: 100%;
   bottom: 0;
@@ -53,29 +50,21 @@
   visibility: hidden;
 }
 
-
-
-@media screen and (max-width: 800px){
+@media screen and (max-width: 800px) {
   .hidden {
     display: none;
   }
   .Dashboard {
-  display: block;
-  padding-top: 10px;
-  padding-bottom: 30px;
-}
-
-.back-router{
-  display:block;
-}
-/* .Dashboard__user-squads,
-.Dashboard__chat {
-  display:none;
-} */
-
-
-  #mobile-buttons-container{
-    visibility: visible;
+    display: block;
+    padding-top: 10px;
+    padding-bottom: 30px;
   }
 
+  .back-router {
+    display: block;
+  }
+
+  #mobile-buttons-container {
+    visibility: visible;
+  }
 }
