+++ src/components/GamesList/GamesList.js
@@ -2,7 +2,6 @@ import React, { useState, useEffect, useContext } from 'react'
 import GameContext from '../../contexts/GameContext'
 import GameApiService from '../../services/game-api.service'
 import GamesSublist from '../GamesSublist/GamesSublist'
-import MobileUtils from '../../components/Utils/MobileUtils'
 import { Input } from '../FormUtils/FormUtils'
 import './GamesList.css'
 
@@ -59,7 +58,6 @@ export default function GamesList() {
 
   return (
     <section className="GamesList">
-    <MobileUtils/>
       <form 
         className="GamesList__search-bar"
       >
